pyproject.toml
[project]
name = "pon-shift-warsaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coverage>=7.10.6",
    "django>=5.2.5",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "fakeredis>=2.31.1",
    "folium>=0.20.0",
    "geopy>=2.4.1",
    "grpcio-tools>=1.75.0",
    "gunicorn>=23.0.0",
    "kombu>=5.5.4",
    "pandas>=2.3.2",
    "polyline>=2.0.3",
    "protobuf>=6.32.1",
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.15.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
]
